Use a newer version of npm
7.6.3

npm 7 is more strict about peer dependency version constraints. I'm
using react 17, but @reach/router does not knowingly work with react 17.
So I had to do this to opt out of the strictness:

```
npm i --legacy-peer-deps
```

I don't really want to do that... so I'm just going to downgrade to
react 16 for now!