Add @shopify/polaris
Needed to update the versions of the react type definitions to match the
ones that polaris uses, because when there are multiple definitions,
typescript isn't happy.

This just plopped in the example from
https://github.com/Shopify/polaris-react#usage

I'll follow up by trying to actually redo the UI using it and see how it
feels.