Leverage query string for search experience
This solves one of the awkward UX problems that your search just lingers
around in the search box forever. Now it will go away once you navigate
away.

It introduces a new UX awkwardness which is that it will be in the URL
bar even before you submit... but that's probably fine.