Fix episode refresh edge case
The night court (2023) refresh is currently failing because, it seems,
they've re-arranged the episode order since I first imported it. That's
fair enough given that they haven't even aired yet.

With this change, we'll rewrite the existing records to reflect whatever
order tmdb says they're in now.

This isn't a totally perfect solution... if they delete an episode after
seasoning has imported it, seasoning won't know to delete it... but
perhaps we'll cross that bridge when we come to it.