Fix: showing flicker of incorrect copy on homepage
If we set loading to false before we set the data then it's possible for
react to redraw the page for a split second and say "No shows yet..."
even though there are and they're going to show up in a split second,
which kinda sucks.

This suggests the modeling of the state for this page is kind of wrong,
but idc, this is a fine quick fix for now.