Improve new hooks scripts
I tried these on a different computer and realized that they don't
ever create the digest file as-written...

I also noticed that this other computer (my fedora linux laptop) has
`md5sum` but not `md5` so I refactored to prefer using md5sum. My Mac
had both. Hopefully it's the same md5sum and it works just as well when
I go back to the Mac and try it there...